applied fixes from M. Laukkanen
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@2761 06827809-a52a-0410-b366-d66718629ded

+++ apps/snmpusm.c
@@ -165,7 +165,7 @@ main(int argc, char *argv[])
     oid                  *authKeyChange   = authKeyOid,
                          *privKeyChange   = privKeyOid;
 
-    u_int                 oldKu_len       = SNMP_MAXBUF_SMALL,
+    size_t                oldKu_len       = SNMP_MAXBUF_SMALL,
                           newKu_len       = SNMP_MAXBUF_SMALL,
                           oldkul_len      = SNMP_MAXBUF_SMALL,
                           newkul_len      = SNMP_MAXBUF_SMALL,
